Hacker Rank - Grid Illumination
Re-creation of solution to "grid illumination" problem based on memory.  Given lamps that illuminate in vertical, horizontal, and diagonal patterns; and given query nodes that deactivate lamps that are adjacent to them, determine if a query node is illuminated after deactivating adjacent lamps.   Input values are given in X,Y coordinates as opposed to standard 2D array notation