Hacker Rank - Pairs
Given an array of integers and a target value, determine the number of pairs of array elements that have a difference equal to the target value.

https://www.hackerrank.com/challenges/pairs/problem

Note: (The solve is straightforward.  The trick was in getting the solve in under the time threshold.  Do not perform unnecessary searches for each possible pairing beforehand. Also bitwise operators work on sets)