Hacker Rank - Largest Permutation
https://www.hackerrank.com/challenges/largest-permutation/problem

Note: (Supposed to be a test on using greedy solutions.  Passes all but last check because artificially implemented speed threshold can only be beat by using a hash.  This implementation will provide a correct answer; will not re-implement because using a hash has nothing to do with solving using a greedy approach)